extraced ToggleButtonBase
